# talde1erronka2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/bezeroak": {
            "get": {
                "tags": [
                    "Bezeroak"
                ],
                "description": "Bezero guztiak bueltatzen ditu.",
                "operationId": "cd5ca9ede919b0fb556d90d11a8beed0",
                "responses": {
                    "200": {
                        "description": "Bezero guztiak bueltatu ditu."
                    },
                    "404": {
                        "description": "Ez du bezerorik topatu."
                    }
                }
            },
            "put": {
                "tags": [
                    "Bezeroak"
                ],
                "description": "Adierazitako informazioaren arabera bezero bat eguneratzen du datubasean.",
                "operationId": "eb6c2d921b1bbe495f9529df81536fd1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Produktuaren ID-a",
                                        "type": "integer"
                                    },
                                    "izena": {
                                        "description": "Produktuaren izena",
                                        "type": "string"
                                    },
                                    "deskribapena": {
                                        "description": "Produktuaren deskribapena",
                                        "type": "string"
                                    },
                                    "id_kategoria": {
                                        "description": "Produktuaren kategoriaren ID-a",
                                        "type": "integer"
                                    },
                                    "marka": {
                                        "description": "Produktuaren marka",
                                        "type": "string"
                                    },
                                    "stock": {
                                        "description": "Produktuaren stock kantitatea",
                                        "type": "integer"
                                    },
                                    "stock_alerta": {
                                        "description": "Alerta aktibatzen duen stock-aren gutxitze-maila",
                                        "type": "integer"
                                    },
                                    "eguneratze_data": {
                                        "description": "Produktuaren eguneratze-data eta ordua",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Bezeroa ondo eguneratu da datubasean."
                    },
                    "404": {
                        "description": "Ez du bezerorik topatu ID horrekin."
                    }
                }
            },
            "post": {
                "tags": [
                    "Bezeroak"
                ],
                "description": "Adierazitako informazioarekin bezero berri bat txertatzen du datubasean.",
                "operationId": "35a470230c0833ba707757ca5547ccda",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "izena": {
                                        "description": "Bezeroaren izena",
                                        "type": "string"
                                    },
                                    "abizena": {
                                        "description": "Bezeroaren abizena",
                                        "type": "string"
                                    },
                                    "telefonoa": {
                                        "description": "Bezeroaren telefono zenbakia",
                                        "type": "string"
                                    },
                                    "piel": {
                                        "description": "Bezeroaren piel (Adibidez, 'E' edo 'H')",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bezeroa ondo txertatu da datubasean."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bezeroak"
                ],
                "description": "Adierazitako ID-a erabilita datubasean bezeroaren ezabatze logikoa egiten du.",
                "operationId": "44a76d7e9146e12c61d22d0846dd8ed6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Elementuaren ID-a",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produktuak ondo ezabatu dira datubasean."
                    },
                    "404": {
                        "description": "Ez dago produkturik ID horrekin."
                    }
                }
            }
        },
        "/bezeroak/{id}": {
            "get": {
                "tags": [
                    "Bezeroak"
                ],
                "description": "Adierazitako ID-aren arabera bezero bat bueltatzen du.",
                "operationId": "9e6b511412725c105566530b4bcaf888",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bezeroaren ID-a",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aukeratutako bezeroa bueltatzen du."
                    },
                    "404": {
                        "description": "Ez du bezerorik topatu ID horrekin."
                    }
                }
            }
        },
        "/devolver": {
            "get": {
                "tags": [
                    "Materiala bueltatu"
                ],
                "description": "Materiala_erabiliko erregistro guztiak bueltatzen ditu.",
                "operationId": "ed7744be569e0750860d436646592bf6",
                "responses": {
                    "200": {
                        "description": "Erregistro guztiak bueltatu ditu."
                    },
                    "404": {
                        "description": "Ez du erregistrorik topatu."
                    }
                }
            },
            "put": {
                "tags": [
                    "Materiala bueltatu"
                ],
                "description": "Adierazitako informazioaren arabera produktu bat eguneratzen du datubasean.",
                "operationId": "f81e4e5b5c029d2012f32827e810b4a6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Jarduera ID-a",
                                        "type": "integer"
                                    },
                                    "id_langilea": {
                                        "description": "Langilearen ID-a",
                                        "type": "integer"
                                    },
                                    "id_materiala": {
                                        "description": "Materialaren ID-a",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Produktua ondo eguneratu da datubasean."
                    },
                    "404": {
                        "description": "Ez du produkturik topatu ID horrekin."
                    }
                }
            },
            "post": {
                "tags": [
                    "Materiala bueltatu"
                ],
                "description": "Adierazitako informazioarekin erregistro berri bat txertatzen du datubasean.",
                "operationId": "bb0e6611ef3ffb4c3876e7fec4244440",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_materiala": {
                                        "description": "Materialaren ID-a",
                                        "type": "integer"
                                    },
                                    "id_langilea": {
                                        "description": "Langilearen ID-a",
                                        "type": "integer"
                                    },
                                    "hasiera_data": {
                                        "description": "Jardueraren hasierako data eta ordua",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "sortze_data": {
                                        "description": "Jardueraren sortze-data eta ordua",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produktua ondo txertatu da datubasean."
                    }
                }
            }
        },
        "/devolver/{id}": {
            "get": {
                "tags": [
                    "Materiala bueltatu"
                ],
                "description": "Adierazitako ID-aren arabera erregistro bat bueltatzen du.",
                "operationId": "f9f5513669dc4dde2af9f1906f973cb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Erregistroaren ID-a",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aukeratutako erregistroa bueltatzen du."
                    },
                    "404": {
                        "description": "Ez du erregistrorik topatu ID horrekin."
                    }
                }
            }
        },
...
    },
...
}